Add check for estab type in fieldwork reminder message Small refactor of step files and feature file names
