+++ acceptance_tests/features/field_reminder.feature
@@ -1,4 +1,4 @@
-Feature: Send messages to fieldwork management tool to keep it in sync with RM
+Feature: Reminder messages are emitted to Field Work Management Tool
 
   Scenario: Tranche 2 household case details to be sent to the Field Work Management Tool
     Given an action rule of type FF2QE is set 15 seconds in the future
