+++ acceptance_tests/features/steps/field_reminder.py
@@ -21,14 +21,6 @@ def fwmt_messages_received(context, treatment_code):
     assert not context.expected_sample_units, 'Some messages are missing'
 
 
-@then('an action instruction cancel message is emitted to FWMT')
-def refusal_received(context):
-    context.seen_expected_fwmt_message = False
-    start_listening_to_rabbit_queue(Config.RABBITMQ_OUTBOUND_FIELD_QUEUE_TEST,
-                                    functools.partial(_refusal_callback, context=context))
-    assert context.seen_expected_fwmt_message, 'Expected message not seen'
-
-
 def _callback(ch, method, _properties, body, context):
     root = ET.fromstring(body)
 
@@ -50,20 +42,6 @@ def _callback(ch, method, _properties, body, context):
         ch.stop_consuming()
 
 
-def _refusal_callback(ch, method, _properties, body, context):
-    root = ET.fromstring(body)
-
-    if not root[0].tag == 'actionCancel':
-        ch.basic_nack(delivery_tag=method.delivery_tag)
-        assert False, 'Unexpected message on Action.Field queue'
-
-    tc.assertEqual(context.refused_case_id, root.find('.//caseId').text)
-    ch.basic_ack(delivery_tag=method.delivery_tag)
-
-    context.seen_expected_fwmt_message = True
-    ch.stop_consuming()
-
-
 def _message_matches(sample_unit, root):
     return root.find('.//arid').text == sample_unit['attributes']['ARID']
 
@@ -75,3 +53,4 @@ def _message_valid(sample_unit, root):
     tc.assertEqual('false', root.find('.//undeliveredAsAddress').text)
     tc.assertEqual('false', root.find('.//blankQreReturned').text)
     tc.assertEqual('CENSUS', root.find('.//surveyName').text)
+    tc.assertEqual(sample_unit['attributes']['ESTAB_TYPE'], root.find('.//estabType').text)
