+++ acceptance_tests/features/steps/refusal.py
@@ -1,9 +1,13 @@
+import functools
 import json
+from xml.etree import ElementTree as ET
 
 import requests
 from behave import step
 
 from acceptance_tests.utilities.rabbit_context import RabbitContext
+from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue
+from acceptance_tests.utilities.test_case_helper import tc
 from config import Config
 
 caseapi_url = f'{Config.CASEAPI_SERVICE}/cases/'
@@ -67,3 +71,25 @@ def check_case_events(context):
         if case_event['description'] == 'Refusal Received':
             return
     assert False
+
+
+@step('an action instruction cancel message is emitted to FWMT')
+def refusal_received(context):
+    context.seen_expected_fwmt_message = False
+    start_listening_to_rabbit_queue(Config.RABBITMQ_OUTBOUND_FIELD_QUEUE_TEST,
+                                    functools.partial(_refusal_callback, context=context))
+    assert context.seen_expected_fwmt_message, 'Expected message not seen'
+
+
+def _refusal_callback(ch, method, _properties, body, context):
+    root = ET.fromstring(body)
+
+    if not root[0].tag == 'actionCancel':
+        ch.basic_nack(delivery_tag=method.delivery_tag)
+        assert False, 'Unexpected message on Action.Field queue'
+
+    tc.assertEqual(context.refused_case_id, root.find('.//caseId').text)
+    ch.basic_ack(delivery_tag=method.delivery_tag)
+
+    context.seen_expected_fwmt_message = True
+    ch.stop_consuming()
